tcr v0.1.0 Tcr

Experiment in Test && Commit || Revert

Installation

git clone https://github.com/mwindholtz/test-commit-revert

In one terminal window run

mix tcr.pull_push This will push commits to the trunk whenever they are ready

In another window run

mix tcr.tcr This will run tests and save commits if test passes, and revert if it does not.

Run the app manually

mix tcr.run --help

References

Kent Beck

JB Rainsberger

Thomas Deniffel

Functions

parse_args(argv)

process(arg1)

run(argv)

Example

$ mix compile; mix tcr.run -h something wonderful